requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyJWT==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.17.0
Werkzeug==2.3.7